Moving evaluateCurrentRoute tests to their own file
